+++ bocadillo/__init__.py
@@ -1,8 +1,10 @@
 from .applications import App
+from .config import settings, SettingsError, configure
 from .compat import ExpectedAsync
 from .errors import HTTPError
 from .middleware import ASGIMiddleware, Middleware
 from .injection import discover_providers, provider, useprovider
+from .plugins import plugin, get_plugins
 from .recipes import Recipe
 from .request import ClientDisconnect, Request
 from .response import Response