+++ bocadillo/sessions.py
@@ -1,2 +1,0 @@
-class MissingSecretKey(Exception):
-    """Raised when no cookie siggnature secret key was found."""