+++ tests/test_allowed_hosts.py
@@ -1,9 +1,9 @@
-from bocadillo import App
+from bocadillo import configure
 from bocadillo.testing import create_client
 
 
-def test_if_host_not_allowed_then_400():
-    app = App(allowed_hosts=["example.com"])
+def test_if_host_not_allowed_then_400(raw_app):
+    app = configure(raw_app, allowed_hosts=["example.com"])
     client = create_client(app)
 
     @app.route("/")