+++ tests/test_cors.py
@@ -1,9 +1,9 @@
-from bocadillo import App
+from bocadillo import configure
 from bocadillo.testing import create_client
 
 
-def test_no_allowed_origins_by_default():
-    app = App(enable_cors=True)
+def test_no_allowed_origins_by_default(raw_app):
+    app = configure(raw_app, cors=True)
 
     @app.route("/")
     async def index(req, res):
@@ -20,8 +20,8 @@ def test_no_allowed_origins_by_default():
     assert response.status_code == 400
 
 
-def test_if_origin_not_in_allow_origins_then_400():
-    app = App(enable_cors=True, cors_config={"allow_origins": ["foobar.com"]})
+def test_if_origin_not_in_allow_origins_then_400(raw_app):
+    app = configure(raw_app, cors={"allow_origins": ["foobar.com"]})
 
     @app.route("/")
     async def index(req, res):
@@ -47,13 +47,10 @@ def test_if_origin_not_in_allow_origins_then_400():
     assert response.status_code == 400
 
 
-def test_if_method_not_in_allow_methods_then_400():
-    app = App(
-        enable_cors=True,
-        cors_config={
-            "allow_origins": ["foobar.com"],
-            "allow_methods": ["POST"],
-        },
+def test_if_method_not_in_allow_methods_then_400(raw_app):
+    app = configure(
+        raw_app,
+        cors={"allow_origins": ["foobar.com"], "allow_methods": ["POST"]},
     )
 
     @app.route("/")