+++ tests/test_gzip.py
@@ -1,9 +1,9 @@
-from bocadillo import App
+from bocadillo import configure
 from bocadillo.testing import create_client
 
 
-def test_if_gzip_enabled_then_response_is_compressed():
-    app = App(enable_gzip=True, gzip_min_size=0)
+def test_if_gzip_enabled_then_response_is_compressed(raw_app):
+    app = configure(raw_app, gzip=True, gzip_min_size=0)
 
     @app.route("/")
     async def index(req, res):