+++ CHANGELOG.md
@@ -16,12 +16,32 @@ As a result, we strongly recommend you read this document carefully before upgra
 
 ### Added
 
+Routing:
+
 - Route parameters are now validated based on type annotations defined on the HTTP or WebSocket view. Annotations can be `int`, `float`, `bool`, `date`, `datetime`, `time`, `decimal.Decimal` or any [TypeSystem] field.
 - Query parameters can be injected into a view by declaring them as parameters with defaults, e.g. `limit: int = None`. Type annotation-based validation is also available.
+
+New settings infrastructure:
+
+- Django-style `bocadillo.settings` lazy object.
+- Configure an app before serving using `bocadillo.configure(app)`.
+
+New plugin mechanism:
+
+- Signature: `(app: App) -> None`.
+- Use the new lazy `settings` object to perform conditional logic.
+- Register a new plugin using `@bocadillo.plugin`.
+- Inspect registered plugins using the `get_plugins()` helper.
+
+Other:
+
 - Error handlers can now re-raise exceptions for further processing, e.g. re-raise an `HTTPError` which will be processed by the registered `HTTPError` handler.
 - Build a full URL for a `LiveServer` using `server.url("/path")`. Note: `server.url` still gives access to the root live server URL.
+  [typesystem]: https://www.encode.io/typesystem
 
-[typesystem]: https://www.encode.io/typesystem
+### Changed
+
+- Features such as CORS, HSTS or allowed hosts are now implemented via plugins, and configured via `app.configure()`. See the "Removed" section for the impact on the application API.
 
 ### Fixed
 
@@ -29,6 +49,14 @@ As a result, we strongly recommend you read this document carefully before upgra
 
 ### Removed
 
+`App` parameters:
+
+- **BREAKING** Removed `static_dir`, `static_root` and `static_config`. Use the `STATIC_DIR`, `STATIC_ROOT` and `STATIC_CONFIG` settings instead.
+- **BREAKING** Removed `allowed_hosts`. Use the `ALLOWED_HOSTS` setting instead.
+- **BREAKING** Removed `enable_sessions` and `sessions_config`. Use the `SESSIONS` setting instead.
+- **BREAKING** Removed `enable_cors` and `cors_config`. Use the `CORS` setting instead.
+- **BREAKING** Removed `enable_hsts`. Use the `HSTS` setting instead.
+- # **BREAKING** Removed `enable_gzip` and `gzip_min_size`. Use the `GZIP` and `GZIP_MIN_SIZE` settings instead.
 - **BREAKING**: the `.run()` method on `App` has been removed in favor of the `uvicorn` command shipped with the [uvicorn] ASGI server (which comes installed with Bocadillo). In particular, the `if __name__ == "__main__": app.run()` invokation is now obsolete. Just use `uvicorn app:app` instead of `python app.py` (and `uvicorn.run(app)` for programmatic usage).
 - **BREAKING**: synchronous views, HTTP middleware callbacks, hooks and error handlers are not supported anymore. Appropriate error messages will help you migrate to an all-async application.
 - **BREAKING**: route parameter validation via specifiers (e.g. `{id:d}`) is not supported anymore. Please use type annotation-based validation instead (e.g. `pk: int`).